Wait for the remote endpoint to close
When a client signals it is about to close the connection by sending
a TPM_SESSION_END command, wait for the client to actually close its
endpoint before closing the local endpoint. This avoids the scenario
where the local endpoint closes first, leaving the connection around
in the kernel in TIME_WAIT mode.